Refactor: Split components.ts into individual component type files
